+++ src/main/java/org/opensearch/security/privileges/PrivilegesEvaluator.java
@@ -76,6 +76,7 @@ import org.opensearch.action.termvectors.MultiTermVectorsAction;
 import org.opensearch.action.update.UpdateAction;
 import org.opensearch.cluster.ClusterState;
 import org.opensearch.cluster.metadata.AliasMetadata;
+import org.opensearch.cluster.metadata.IndexAbstraction;
 import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.metadata.Metadata;
@@ -93,6 +94,7 @@ import org.opensearch.security.configuration.ConfigurationRepository;
 import org.opensearch.security.resolver.IndexResolverReplacer;
 import org.opensearch.security.resolver.IndexResolverReplacer.Resolved;
 import org.opensearch.security.securityconf.ConfigModel;
+import org.opensearch.security.securityconf.DynamicConfigFactory;
 import org.opensearch.security.securityconf.DynamicConfigModel;
 import org.opensearch.security.securityconf.FlattenedActionGroups;
 import org.opensearch.security.securityconf.impl.CType;
@@ -155,9 +157,18 @@ public class PrivilegesEvaluator {
     private DynamicConfigModel dcm;
     private final NamedXContentRegistry namedXContentRegistry;
     private final Settings settings;
-    private final Map<String, Set<String>> pluginToClusterActions;
-    private final AtomicReference<ActionPrivileges> actionPrivileges = new AtomicReference<>();
+    private final AtomicReference<RoleBasedActionPrivileges> actionPrivileges = new AtomicReference<>();
     private final AtomicReference<TenantPrivileges> tenantPrivileges = new AtomicReference<>();
+    private final Map<String, SubjectBasedActionPrivileges> pluginIdToActionPrivileges = new HashMap<>();
+    private final Supplier<Map<String, IndexAbstraction>> indexMetadataSupplier;
+
+    /**
+     * The pure static action groups should be ONLY used by action privileges for plugins; only those cannot and should
+     * not have knowledge of any action groups defined in the dynamic configuration. All other functionality should
+     * use the action groups derived from the dynamic configuration (which is always computed on the fly on
+     * configuration updates).
+     */
+    private final FlattenedActionGroups staticActionGroups;
 
     public PrivilegesEvaluator(
         final ClusterService clusterService,
@@ -180,7 +191,6 @@ public class PrivilegesEvaluator {
 
         this.threadContext = threadContext;
         this.privilegesInterceptor = privilegesInterceptor;
-        this.pluginToClusterActions = new HashMap<>();
         this.clusterStateSupplier = clusterStateSupplier;
         this.settings = settings;
 
@@ -198,6 +208,10 @@ public class PrivilegesEvaluator {
         pitPrivilegesEvaluator = new PitPrivilegesEvaluator();
         this.namedXContentRegistry = namedXContentRegistry;
         this.configurationRepository = configurationRepository;
+        this.staticActionGroups = new FlattenedActionGroups(
+            DynamicConfigFactory.addStatics(SecurityDynamicConfiguration.empty(CType.ACTIONGROUPS))
+        );
+        this.indexMetadataSupplier = () -> clusterStateSupplier.get().metadata().getIndicesLookup();
 
         if (configurationRepository != null) {
             configurationRepository.subscribeOnChange(configMap -> {
@@ -213,7 +227,7 @@ public class PrivilegesEvaluator {
 
         if (clusterService != null) {
             clusterService.addListener(event -> {
-                ActionPrivileges actionPrivileges = PrivilegesEvaluator.this.actionPrivileges.get();
+                RoleBasedActionPrivileges actionPrivileges = PrivilegesEvaluator.this.actionPrivileges.get();
                 if (actionPrivileges != null) {
                     actionPrivileges.updateClusterStateMetadataAsync(clusterService, threadPool);
                 }
@@ -231,16 +245,15 @@ public class PrivilegesEvaluator {
         rolesConfiguration = rolesConfiguration.withStaticConfig();
         tenantConfiguration = tenantConfiguration.withStaticConfig();
         try {
-            ActionPrivileges actionPrivileges = new ActionPrivileges(
+            RoleBasedActionPrivileges actionPrivileges = new RoleBasedActionPrivileges(
                 rolesConfiguration,
                 flattenedActionGroups,
-                () -> clusterStateSupplier.get().metadata().getIndicesLookup(),
-                settings,
-                pluginToClusterActions
+                indexMetadataSupplier,
+                settings
             );
             Metadata metadata = clusterStateSupplier.get().metadata();
             actionPrivileges.updateStatefulIndexPrivileges(metadata.getIndicesLookup(), metadata.version());
-            ActionPrivileges oldInstance = this.actionPrivileges.getAndSet(actionPrivileges);
+            RoleBasedActionPrivileges oldInstance = this.actionPrivileges.getAndSet(actionPrivileges);
 
             if (oldInstance != null) {
                 oldInstance.shutdown();
@@ -266,13 +279,9 @@ public class PrivilegesEvaluator {
         this.dcm = dcm;
     }
 
-    public ActionPrivileges getActionPrivileges() {
-        return this.actionPrivileges.get();
-    }
-
     public boolean hasRestAdminPermissions(final User user, final TransportAddress remoteAddress, final String permission) {
         PrivilegesEvaluationContext context = createContext(user, permission);
-        return this.actionPrivileges.get().hasExplicitClusterPrivilege(context, permission).isAllowed();
+        return context.getActionPrivileges().hasExplicitClusterPrivilege(context, permission).isAllowed();
     }
 
     public boolean isInitialized() {
@@ -311,9 +320,32 @@ public class PrivilegesEvaluator {
         }
 
         TransportAddress caller = threadContext.getTransient(ConfigConstants.OPENDISTRO_SECURITY_REMOTE_ADDRESS);
-        ImmutableSet<String> mappedRoles = ImmutableSet.copyOf((injectedRoles == null) ? mapRoles(user, caller) : injectedRoles);
 
-        return new PrivilegesEvaluationContext(user, mappedRoles, action0, request, task, irr, resolver, clusterStateSupplier);
+        ActionPrivileges actionPrivileges;
+        ImmutableSet<String> mappedRoles;
+
+        if (user.isPluginUser()) {
+            mappedRoles = ImmutableSet.of();
+            actionPrivileges = this.pluginIdToActionPrivileges.get(user.getName());
+            if (actionPrivileges == null) {
+                actionPrivileges = ActionPrivileges.EMPTY;
+            }
+        } else {
+            mappedRoles = ImmutableSet.copyOf((injectedRoles == null) ? mapRoles(user, caller) : injectedRoles);
+            actionPrivileges = this.actionPrivileges.get();
+        }
+
+        return new PrivilegesEvaluationContext(
+            user,
+            mappedRoles,
+            action0,
+            request,
+            task,
+            irr,
+            resolver,
+            clusterStateSupplier,
+            actionPrivileges
+        );
     }
 
     public PrivilegesEvaluatorResponse evaluate(PrivilegesEvaluationContext context) {
@@ -366,7 +398,7 @@ public class PrivilegesEvaluator {
             log.debug("Mapped roles: {}", mappedRoles.toString());
         }
 
-        ActionPrivileges actionPrivileges = this.actionPrivileges.get();
+        ActionPrivileges actionPrivileges = context.getActionPrivileges();
         if (actionPrivileges == null) {
             throw new OpenSearchSecurityException("OpenSearch Security is not initialized: roles configuration is missing");
         }
@@ -851,5 +883,10 @@ public class PrivilegesEvaluator {
 
     public void updatePluginToClusterActions(String pluginIdentifier, Set<String> clusterActions) {
         pluginToClusterActions.put(pluginIdentifier, clusterActions);
+    public void updatePluginToPermissions(String pluginIdentifier, RoleV7 pluginPermissions) {
+        this.pluginIdToActionPrivileges.put(
+            pluginIdentifier,
+            new SubjectBasedActionPrivileges(pluginPermissions, this.staticActionGroups, this.indexMetadataSupplier)
+        );
     }
 }
